Totem Boss tuning
Totem Speaker Boss now has 50 HP.

Totems reduced to 45 HP.

Totems that spawn in after the initial 3 now show the 'Stun' intent on their first turn, clarifying they will not attack.

Totems that are Stunned will not counterattack on AoE cards until the following turn.  This is visually shown with a greyed out Counterattack buff icon and an alternate tooltip for the buff effect.

Totems which apply Block no longer apply Block to themselves, only to the other living totems.