Totem Boss clarification visuals and tuning
Totems reduced to 40 HP.

When the fight begins, Totems now show a visual effect powering up the Totem Speaker with its Immunity buff.  The immunity buff now shows a stack count indicating how many totems remain in the fight.  These should help clarify that there are 6 totems to kill in total.

Totem Speaker now applies Frail and Weak on the debuff turn, not Vulnerable and Weak.  This will make the fight less spiky in terms of damage, as the Vulnerable turn was always applied the turn before the x6 attack.